Canary - Filters, Labels with a single non-boolean value don't show the value
